+++ FirebaseContext/firebase.js
@@ -17,6 +17,8 @@ class Firebase {
 
     queryUsers = () => this.db.collection("users")
 
+    queryOneUsers = (id) => this.db.collection("users").doc(id).get();
+
     //add un user
 
     queryAddUser = (id,data) => firestore.collection("users").doc(id).set(data);
@@ -41,7 +43,7 @@ class Firebase {
     
     //lire comment
     
-    queryDetails = () => this.db.collection("comment").orderBy('date','desc');
+    queryDetails = (id) => this.db.collection("comment").where('gallery','==',id).orderBy('date','desc');
 
     queryDetailId = (id) => this.db.collection("comment").doc(id).get();
